+++ CHANGELOG.md
@@ -101,3 +101,4 @@
    - If a DAL returns None for a number looking type return 0 instead.
    - Deprecated TemplateNinja
    - Fix for matching uint/decimal64 in a union (the dal returns a string not a python int)
+   - Add `get_raw_xpath` to sublydal